move MintableToken.sol to ERC20Mintable.sol, likewise test and mock files
