+++ contracts/crowdsale/emission/MintedCrowdsale.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.4.24;
 
 import "../Crowdsale.sol";
-import "../../token/ERC20/MintableToken.sol";
+import "../../token/ERC20/ERC20Mintable.sol";
 
 
 /**
