+++ contracts/mocks/FinalizableCrowdsaleImpl.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.4.24;
 
-import "../token/ERC20/MintableToken.sol";
+import "../token/ERC20/ERC20Mintable.sol";
 import "../crowdsale/distribution/FinalizableCrowdsale.sol";
