+++ test/token/ERC20/ERC20Mintable.behavior.js
