+++ test/token/ERC20/RBACMintableToken.test.js
@@ -1,5 +1,5 @@
 const { shouldBehaveLikeRBACMintableToken } = require('./RBACMintableToken.behavior');
-const { shouldBehaveLikeERC20Mintable } = require('./MintableToken.behavior');
+const { shouldBehaveLikeERC20Mintable } = require('./ERC20Mintable.behavior');
 
 const RBACMintableToken = artifacts.require('RBACMintableToken');
